backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
opencv-python==4.8.1.78
mediapipe==0.10.8
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.1.0
